requirements.txt
streamlit
langchain
tabulate
python-dotenv
